refactoring(lab5): добавлена возможность использовать сертификат при оплате нескольких товаров

+++ lab5/lab5/CertificateDiscount.h
@@ -9,7 +9,7 @@ private:
 public:
 	void SetAmount(const float amount);
 
-	double GetAmount();
+	float GetAmount();
 
 	CertificateDiscount(const float amount, 
 		const CategoryType& category);
